backend/requirements.txt
fastapi
uvicorn
python-multipart
pydicom
opencv-python
Pillow
requests
python-dotenv